requirements.txt
discord.py>=1.2.5
jthon>=1.0.2